% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeforSSR_q2.2.1.R
\name{bayespredpower_wh}
\alias{bayespredpower_wh}
\title{Bayesian predictive power calculation using  Wilson-Hilferty transformation}
\usage{
bayespredpower_wh(psi1, n2, n, alphalevel)
}
\arguments{
\item{psi1}{the calculated CMH test statistics from the CMH test at interim analysis.}

\item{n2}{the initial planning total sample size of two arms at stage 2.}

\item{n}{the total sample size of two arms at stage 1.}

\item{alphalevel}{the alpha level of the one-side test.}
}
\description{
This function calculates the bayesian predictive power under CMH test using non-informative prior and Wilson-Hilferty transformation as in formula (24) and (25).
This function will be called for in the ssr_bpp_wh function.
}
\examples{
library(SSreEstimate)
bpp_wh <- bayespredpower_wh(4,100,100,0.025)

}
