% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeforSSR_q2.2.1.R
\name{conditionalpower_wh}
\alias{conditionalpower_wh}
\title{Conditional power calculation using Wilson-Hilferty transformation}
\usage{
conditionalpower_wh(psi1, n2, n, alphalevel)
}
\arguments{
\item{psi1}{the calculated CMH test statistics from the CMH test at interim analysis.}

\item{n2}{the initial planning total sample size of two arms at stage 2.}

\item{n}{the total sample size of two arms at stage 1.}

\item{alphalevel}{the alpha level of the one-side test.}
}
\description{
This function calculates the conditional power under CMH test using Wilson-Hilferty transformation  as in formula (14) and (15).
This function will be called for in the ssr_cp_wh function.
}
\examples{
library(SSreEstimate)
cp_wh <- conditionalpower_wh(4,100,100,0.025)

}
