% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeforSSR_q2.2.1.R
\name{ssr_cp}
\alias{ssr_cp}
\title{Sample size re-estimation based on conditional power}
\usage{
ssr_cp(allo, n2plan, n2max, nnow, ptnow, pcnow, alpha, delta, target_power)
}
\arguments{
\item{allo}{the allocation rate into the treatment arm.}

\item{n2plan}{the initial planned sample size of two arms for stage2.}

\item{n2max}{the allowed largest sample size considering the medical resource.}

\item{nnow}{the total sample size of two arms at stage 1.}

\item{ptnow}{the calculated response rate in treatment arm at interim analysis.}

\item{pcnow}{the calculated response rate in control arm at interim analysis.}

\item{alpha}{the alpha level of the one-side test.}

\item{delta}{the difference of efficacy expected to be detected in the hyphothesises, usually 0.}

\item{target_power}{the target power wish to achieve.}
}
\description{
This function is for sample size re-estimation based on conditional power as in formula (11).
This function is based on the conditionalpower function.
}
\examples{
 library(SSreEstimate)
 ssr_cp(1,100,300,100,0.55,0.4,0.025,0,0.8)

}
