% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeforSSR_q2.2.1.R
\name{ssr_cp_wh}
\alias{ssr_cp_wh}
\title{Sample size re-estimation based on conditional power using  Wilson-Hilferty transformation}
\usage{
ssr_cp_wh(psi1value, n2plan, n2max, nnow, alpha, target_power)
}
\arguments{
\item{psi1value}{the calculated CMH test statistics from the CMH test at interim analysis.}

\item{n2plan}{the initial planning total sample size of two arms at stage 2.}

\item{n2max}{the allowed largest sample size considering the medical resource.}

\item{nnow}{the total sample size of two arms at stage 1.}

\item{alpha}{the alpha level of the one-side test.}

\item{target_power}{the target power wish to achieve.}
}
\description{
This function is for sample size re-estimation with CMH test based on conditional power and Wilson-Hilferty transformation as in formula (16).
This function is based on the conditionalpower_wh function.
}
\examples{
library(SSreEstimate)
ssr_cp_wh(5,100,300,200,0.025,0.8)

}
